Final clean version of upload files
